use tree bandit as default for now fix bounds on learner_fetch Setup improvement for game_gen (skip pass turns) fix value setting to actor buffer, finish actions policy indices fix copy_game_to_sample, wrong dtype size on cpy fix thread count on sample kernel

+++ src/common.hh
@@ -12,7 +12,7 @@ struct ActorBuffers
     uint64_t *raw_input_buffer;
     float *value_data_buffer;
     float *joined_policy_buffer;
-    uint32_t *joined_policy_index_buffer;
+    int64_t *joined_policy_index_buffer;
 };
 
 struct LearnerBuffers
@@ -20,7 +20,7 @@ struct LearnerBuffers
     float *float_input_buffer;
     float *value_data_buffer;
     float *joined_policy_buffer;
-    uint32_t *joined_policy_index_buffer;
+    int64_t *joined_policy_index_buffer;
 };
 
 namespace CUDACommon
